+++ packages/radix-vue/src/Select/index.ts
@@ -3,12 +3,10 @@ export { default as SelectTrigger } from "./SelectTrigger.vue";
 export { default as SelectPortal } from "./SelectPortal.vue";
 export { default as SelectContent } from "./SelectContent.vue";
 export { default as SelectArrow } from "./SelectArrow.vue";
-export { default as SelectGroup } from "./SelectGroup.vue";
 export { default as SelectSeparator } from "./SelectSeparator.vue";
-export { default as SelectCheckboxItem } from "./SelectCheckboxItem.vue";
 export { default as SelectItemIndicator } from "./SelectItemIndicator.vue";
 export { default as SelectLabel } from "./SelectLabel.vue";
-export { default as SelectRadioGroup } from "./SelectRadioGroup.vue";
-export { default as SelectRadioItem } from "./SelectRadioItem.vue";
+export { default as SelectGroup } from "./SelectGroup.vue";
+export { default as SelectItem } from "./SelectItem.vue";
 
 export * from "./utils";
